+++ BTDB/IL/Caching/CachingILDynamicType.cs
@@ -279,7 +279,12 @@ namespace BTDB.IL.Caching
 
         public IILMethod DefineConstructor(Type[] parameters)
         {
-            var res = new Constructor((int)_instructions.Count, parameters);
+            return DefineConstructor(parameters, Array.Empty<string>());
+        }
+
+        public IILMethod DefineConstructor(Type[] parameters, string[] parametersNames)
+        {
+            var res = new Constructor((int)_instructions.Count, parameters, parametersNames);
             _instructions.Add(res);
             return res;
         }
@@ -288,20 +293,22 @@ namespace BTDB.IL.Caching
         {
             readonly int _id;
             readonly Type[] _parameters;
+            readonly string[] _parametersNames;
             readonly CachingILGen _ilGen = new CachingILGen();
             int _expectedLength = -1;
             IILMethod? _trueContent;
 
-            public Constructor(int id, Type[] parameters)
+            public Constructor(int id, Type[] parameters, string[] parametersNames)
             {
                 _id = id;
                 _parameters = parameters;
+                _parametersNames = parametersNames;
                 InitLocals = true;
             }
 
             public void ReplayTo(IILDynamicType target)
             {
-                _trueContent = target.DefineConstructor(_parameters);
+                _trueContent = target.DefineConstructor(_parameters, _parametersNames);
                 _trueContent.InitLocals = InitLocals;
                 if (_expectedLength >= 0) _trueContent.ExpectedLength(_expectedLength);
             }
@@ -320,7 +327,7 @@ namespace BTDB.IL.Caching
             public bool Equals(IReplay? other)
             {
                 if (!(other is Constructor v)) return false;
-                return _id == v._id && _parameters.SequenceEqual(v._parameters) && _ilGen.Equals(v._ilGen) && InitLocals == v.InitLocals;
+                return _id == v._id && _parameters.SequenceEqual(v._parameters) && _ilGen.Equals(v._ilGen) && _parametersNames.SequenceEqual(v._parametersNames) && InitLocals == v.InitLocals;
             }
 
             public override bool Equals(object obj)