+++ BTDBTest/EventStore2Test.cs
@@ -1018,5 +1018,111 @@ namespace BTDBTest
             Store(new Dictionary<int, IList<bool>> { { 1, new List<bool> { true } } });
             Store(new Dictionary<int, IList<bool>> { { 1, new[] { true } } });
         }
+
+        [Fact]
+        public void DictionaryWithSomeNullListAsValue()
+        {
+            var serializer = new EventSerializer();
+            var obj = new Dictionary<ulong, IList<ulong>>
+            {
+                {1, null},
+                {2, new List<ulong> {21, 22}},
+                {3, null},
+                {4, new List<ulong> {41, 42}}
+            };
+            var meta = serializer.Serialize(out _, obj).ToAsyncSafe();
+            serializer.ProcessMetadataLog(meta);
+            var data = serializer.Serialize(out _, obj);
+
+            var deserializer = new EventDeserializer();
+            Assert.False(deserializer.Deserialize(out var obj2, data));
+            deserializer.ProcessMetadataLog(meta);
+            Assert.True(deserializer.Deserialize(out obj2, data));
+
+            Assert.Equal(obj, obj2);
+        }
+
+        [Fact]
+        public void DictionaryWithSomeNullArrayAsValue()
+        {
+            var serializer = new EventSerializer();
+            var obj = new Dictionary<ulong, ulong[]>
+            {
+                {1, null},
+                {2, new ulong[] {21, 22}},
+                {3, null},
+                {4, new ulong[] {41, 42}}
+            };
+            var meta = serializer.Serialize(out _, obj).ToAsyncSafe();
+            serializer.ProcessMetadataLog(meta);
+            var data = serializer.Serialize(out _, obj);
+
+            var deserializer = new EventDeserializer();
+            Assert.False(deserializer.Deserialize(out var obj2, data));
+            deserializer.ProcessMetadataLog(meta);
+            Assert.True(deserializer.Deserialize(out obj2, data));
+
+            Assert.Equal(obj, obj2);
+        }
+
+        [Fact(Skip = "Generic serialization of structs is hard to implement, that's why it is not working for now.")]
+        public void SerializationOfStruct_Succeeds()
+        {
+            var testStruct = new TestStruct { TestData = new TestStruct.TestStructData { Data = "TestData" } };
+
+            var result = SerializationInternal<TestStruct>(testStruct);
+
+            Assert.NotNull(result);
+            Assert.Equal(testStruct.TestData.Data, result.TestData.Data);
+        }
+
+        [Fact]
+        public void SerializationOfBaseClassWithPrivateSet_Succeeds()
+        {
+            var testBaseClass = new TestClassWithBaseClass();
+            var result = SerializationInternal<TestClassWithBaseClass>(testBaseClass);
+
+            Assert.NotNull(result);
+            Assert.Equal(testBaseClass.TestData, result.TestData);
+        }
+
+        private T SerializationInternal<T>(object input)
+        {
+            var serializer = new EventSerializer();
+            var meta = serializer.Serialize(out _, input).ToAsyncSafe();
+            serializer.ProcessMetadataLog(meta);
+            var data = serializer.Serialize(out _, input);
+
+            var deserializer = new EventDeserializer();
+            deserializer.ProcessMetadataLog(meta);
+            deserializer.Deserialize(out var deserializedObj, data);
+
+            return (T)deserializedObj;
+        }
+
+        class TestStruct
+        {
+            public TestStructData TestData { get; set; }
+
+            internal struct TestStructData
+            {
+                public string Data { get; set; }
+            }
+        }
+
+        class TestClassWithBaseClass : TestBaseClass
+        {
+            [NotStored]
+            public string TestData => Data;
+        }
+
+        class TestBaseClass
+        {
+            public string Data { get; private set; }
+            public TestBaseClass()
+            {
+                Data = "TestData";
+            }
+        }
     }
 }